requirements.txt
apispec==6.9.0
blinker==1.9.0
click==8.3.1
et_xmlfile==2.0.0
Flask==3.1.2
flask-smorest==0.46.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.1.2
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.2
webargs==8.7.1
Werkzeug==3.1.4
xlsxwriter==3.2.9
apispec==6.9.0
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
et_xmlfile==2.0.0
Flask==3.1.2
Flask-JWT-Extended==4.7.1
flask-smorest==0.46.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.1.2
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.2
webargs==8.7.1
Werkzeug==3.1.4
xlsxwriter==3.2.9
